Move test_joining to reshaping directory
